Divide classes and separate to smaller files
