+++ WultraMobileTokenSDK/Operations/Model/UserOperation/Screens/PreApproval/Elements/WMTPreApprovalElementButton.swift
@@ -1,0 +1,46 @@
+//
+// Copyright 2025 Wultra s.r.o.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions
+// and limitations under the License.
+//
+
+import Foundation
+
+/// Button element (action + optional href + text).
+public final class WMTPreApprovalElementButton: WMTPreApprovalElement {
+
+    /// Supported button actions.
+    public enum ButtonAction: String, Codable { case link = "LINK", mail = "MAIL", phone = "PHONE", reject = "REJECT" }
+
+    /// Action for the button.
+    public let action: ButtonAction?
+
+    /// URL / resource reference (for `LINK`, `MAIL`, `PHONE` actions).
+    public let href: String?
+
+    private enum Keys: String, CodingKey { case action, href }
+
+    /// Public convenience initializer.
+    public init(id: String? = nil, action: ButtonAction? = nil, href: String? = nil, icon: String? = nil, text: String? = nil) {
+        self.action = action
+        self.href = href
+        super.init(id: id, type: .button, icon: icon, text: text)
+    }
+
+    public required init(from decoder: Decoder) throws {
+        let c = try decoder.container(keyedBy: Keys.self)
+        self.action = try? c.decode(ButtonAction.self, forKey: .action)
+        self.href   = try? c.decode(String.self, forKey: .href)
+        try super.init(from: decoder)
+    }
+}
